Infinity ScanF Commit 1
Created a useful function called "InfinityScanF", to take input from the user and store it in an array. Array size is not need be known since the array is dynamically grown as a new character is encountered, returning the pointer to it, ready to be captured. This does require other libraries I've created.